Вспомогательные функции, необходимые для рендеринга компонентов перенесены из файла render.js в utils.js. Файл render.js удалён.
