+++ src/main.js
@@ -1,6 +1,6 @@
 import FiltersFormView from './view/filters_form/filters-form-view.js';
 import RoutePresenter from './presenter/route-presenter.js';
-import {render} from './render.js';
+import {render} from './utils.js';
 import PointsModel from './model/points-model.js';
 import OffersModel from './model/offers-model.js';
