+++ src/utils.js
@@ -12,4 +12,38 @@ const getRandomInteger = (a = 0, b = 1) => {
 //Функция преобразующая дату из формата '2022-04-29T15:59:39.600Z' в формат 'Apr 29'
 const humanizeDateMonthDay = (anyDate) => dayjs(anyDate).format('MMM DD');
 
-export {getRandomInteger, humanizeDateMonthDay};
+//Фукция для рендеринга передаваемых в неё компонентов (шаблонов)
+const RenderPosition = {
+  BEFOREBEGIN: 'beforebegin',
+  AFTERBEGIN: 'afterbegin',
+  BEFOREEND: 'beforeend',
+  AFTEREND: 'afterend',
+};
+
+const createElement = (template) => {
+  const newElement = document.createElement('div');
+  newElement.innerHTML = template;
+
+  return newElement.firstElementChild;
+};
+
+const render = (component, container, place = RenderPosition.BEFOREEND) => {
+  const element = component.getElement();
+
+  switch (place) {
+    case RenderPosition.BEFOREBEGIN:
+      container.before(element);
+      break;
+    case RenderPosition.AFTERBEGIN:
+      container.prepend(element);
+      break;
+    case RenderPosition.BEFOREEND:
+      container.append(element);
+      break;
+    case RenderPosition.AFTEREND:
+      container.after(element);
+      break;
+  }
+};
+
+export {getRandomInteger, humanizeDateMonthDay, RenderPosition, createElement, render};
