+++ src/view/edit_event_form/edit-event-form-view.js
@@ -1,4 +1,4 @@
-import {createElement} from '../../render.js';
+import {createElement} from '../../utils.js';
 import {createEditEventFormTemplate} from './edit-event-form.tpl.js';
 
 export default class EditEventFormView {
