+++ src/view/sort_form/sort-form-view.js
@@ -1,4 +1,4 @@
-import {createElement} from '../../render.js';
+import {createElement} from '../../utils.js';
 import {createSortFormTemplate} from './sort-form.tpl.js';
 
 export default class SortFormView {
